Refactoring the basic game objects.
